move lambda file
